Extract the analytics panel from the game detail dialog.
